% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_ARACNe}
\alias{add_ARACNe}
\title{Add ARACNe edges to an existing igraph network}
\usage{
add_ARACNe(reg_net, aracne, MI_threshold = NULL)
}
\arguments{
\item{reg_net}{An \code{igraph} object.}

\item{aracne}{A data frame.}

\item{MI_threshold}{Numeric.}
}
\value{
An \code{igraph} object.
}
\description{
Add ARACNe edges to an existing igraph network
}
