% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_ARACNe_by_MI}
\alias{add_ARACNe_by_MI}
\title{Add ARACNe edges to a regulatory network using MI threshold}
\usage{
add_ARACNe_by_MI(reg_net, dta, MI_threshold = NULL)
}
\arguments{
\item{reg_net}{An igraph object.}

\item{dta}{A list.}

\item{MI_threshold}{Numeric or NULL.}
}
\value{
An igraph object.
}
\description{
Add ARACNe edges to a regulatory network using MI threshold
}
