% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_node_attributes}
\alias{add_node_attributes}
\title{Add node attributes to an igraph object}
\usage{
add_node_attributes(
  ig,
  nodes,
  node_name = "gene",
  rm_miss = TRUE,
  reverse_check = TRUE
)
}
\arguments{
\item{ig}{An \code{igraph} object. Vertices must have a \code{name} attribute.}

\item{nodes}{A data frame or tibble containing node attributes. One column must match
the vertex names in the graph (default is \code{"gene"}).}

\item{node_name}{Character string. Column name in \code{nodes} used to match graph vertices (default = \code{"gene"}).}

\item{rm_miss}{Logical. If \code{TRUE}, remove vertices in the graph not present in \code{nodes} (default = \code{TRUE}).}

\item{reverse_check}{Logical. If \code{TRUE}, filter \code{nodes} to keep only rows present in the graph (default = \code{TRUE}).}
}
\value{
An \code{igraph} object with updated vertex attributes.
}
\description{
This function adds attributes from a data frame to the vertices of an \code{igraph} object.
Optionally, it can remove vertices not present in the data frame and ensure only nodes
present in the graph are used for assignment.
}
