% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{add_STRINGdb_edges}
\alias{add_STRINGdb_edges}
\title{Add STRINGdb Edges to an igraph Object}
\usage{
add_STRINGdb_edges(ig, stringdb_obj, simplify = FALSE)
}
\arguments{
\item{ig}{An igraph object representing the existing network.}

\item{stringdb_obj}{A STRINGdb object.}

\item{simplify}{Logical, if TRUE, remove duplicated edges and loops (default = FALSE).}
}
\value{
An igraph object.
}
\description{
This function takes an igraph object and a STRINGdb object, retrieves the edges from STRINGdb
that correspond to the nodes in the igraph object, and combines them with the existing edges
in the igraph object.
}
