% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calc_fdr}
\alias{calc_fdr}
\title{Calculate FDR (False Discovery Rate) from p-values}
\usage{
calc_fdr(df, pval.col, by.col = NULL)
}
\arguments{
\item{df}{Data frame of p-values.}

\item{pval.col}{P-value column name.}

\item{by.col}{Grouping columns.}
}
\value{
Data frame with fdr column.
}
\description{
Calculate FDR (False Discovery Rate) from p-values
}
