% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{combine_jeme_hic}
\alias{combine_jeme_hic}
\title{Combine JEME and HiC networks into a single regulatory network}
\usage{
combine_jeme_hic(
  jeme_harmonized,
  hic_harmonized,
  harmonized = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{jeme_harmonized}{Output list from \code{harmonize_jeme_genecode()}}

\item{hic_harmonized}{Output list from \code{harmonize_hic_gtf()}}

\item{harmonized}{Logical, whether to use harmonized gene names (default = TRUE)}

\item{verbose}{Logical, whether to print progress messages (default = TRUE)}
}
\value{
A tibble with columns \code{Regulator}, \code{Target}, and \code{source}
}
\description{
This function combines harmonized JEME and HiC (PO and PP) networks
into a single data frame with regulator-target pairs and a source column.
}
