% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_GTEx.R
\name{create_aracne_slurm}
\alias{create_aracne_slurm}
\title{Create a SLURM script for ARACNe analysis}
\usage{
create_aracne_slurm(
  tissue = "Esophagus_Gastroesophageal_Junction",
  output_dir = ".",
  aracne_jar = "../ARACNe/aracne.jar",
  partition = "rpci",
  cpus = 32,
  mem = "64G",
  bootstrap = 500
)
}
\arguments{
\item{tissue}{Tissue name.}

\item{output_dir}{Directory where the work is performed.}

\item{aracne_jar}{Path to the aracne.jar file (relative to the output_dir or absolute).}

\item{partition}{SLURM partition (default "rpci").}

\item{cpus}{SLURM cpus-per-task (default 32).}

\item{mem}{SLURM memory (default "64G").}

\item{bootstrap}{Number of bootstraps (default 500).}
}
\value{
Path to the created SLURM script (invisibly).
}
\description{
Create a SLURM script for ARACNe analysis
}
