% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_igraph_hotnet}
\alias{create_igraph_hotnet}
\title{Create igraph object from hotnet results with cluster-based node coloring and shapes}
\usage{
create_igraph_hotnet(edge_tbl, nodes_tbl, hotnet)
}
\arguments{
\item{edge_tbl}{Data frame of edges.}

\item{nodes_tbl}{Data frame of nodes.}

\item{hotnet}{Hotnet results.}
}
\value{
An igraph object.
}
\description{
Create igraph object from hotnet results with cluster-based node coloring and shapes
}
