% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_GTEx.R
\name{download_gtex_gencode}
\alias{download_gtex_gencode}
\title{Download GENCODE annotation for GTEx}
\usage{
download_gtex_gencode(cache_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{cache_dir}{Directory to cache the downloaded file. If NULL, proposes .cache/ in R project or current dir.}

\item{overwrite}{Logical, whether to overwrite existing file without prompting (default FALSE)}
}
\value{
Path to the downloaded file
}
\description{
Download GENCODE annotation for GTEx
}
