% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_GTEx.R
\name{download_hgnc_complete}
\alias{download_hgnc_complete}
\title{Download HGNC complete set}
\usage{
download_hgnc_complete(cache_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{cache_dir}{Directory to cache the downloaded file. If NULL, proposes .cache/ in R project or current dir.}

\item{overwrite}{Logical, whether to overwrite existing file without prompting (default FALSE)}
}
\value{
Path to the downloaded file
}
\description{
Download HGNC complete set
}
