% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_HiC.R
\name{download_HiC}
\alias{download_HiC}
\title{Download and process HiC data}
\usage{
download_HiC(cache_dir = NULL, out_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{cache_dir}{Directory to save downloaded Excel files; NULL uses the package cache setting.}

\item{out_dir}{Directory to save processed files. Defaults to ".cache/HiC" safely found within the project root.}

\item{overwrite}{Logical, whether to overwrite existing downloaded files (default FALSE).}
}
\value{
A list containing the P-O and P-P interaction dataframes
}
\description{
Download and process HiC data
}
