% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_JEME.R
\name{download_JEME}
\alias{download_JEME}
\title{Download and process JEME data}
\usage{
download_JEME(
  type = c("lasso", "elasticnet"),
  cache_dir = NULL,
  out_dir = NULL,
  save_summarized = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{type}{Type of data to download: "lasso" or "elasticnet"}

\item{cache_dir}{Directory to save downloaded zip files; NULL uses the package cache setting.}

\item{out_dir}{Directory to save processed files. Defaults to ".cache/HiC" safely found within the project root.}

\item{save_summarized}{Logical, whether to create and save summarized enhancer/promoter RDS and TSV files (default TRUE)}

\item{overwrite}{Logical, whether to overwrite existing downloaded zip files (default FALSE).}
}
\value{
A list containing the combined JEME dataframe, unique enhancers, and unique promoters
}
\description{
Download and process JEME data
}
