% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{downloadAbsentFile}
\alias{downloadAbsentFile}
\title{Download a file when it is not already present}
\usage{
downloadAbsentFile(urlStr, oD = tempdir())
}
\description{
Download a file when it is not already present
}
\keyword{internal}
