% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_gene_network_ldak}
\alias{filter_gene_network_ldak}
\title{Filter network edges based on genes present in LDAK scores}
\usage{
filter_gene_network_ldak(net, ldak)
}
\arguments{
\item{net}{Network data frame.}

\item{ldak}{LDAK score data frame.}
}
\value{
Filtered network data frame.
}
\description{
Filter network edges based on genes present in LDAK scores
}
