% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{filter_igraph_hotnet}
\alias{filter_igraph_hotnet}
\title{Filter igraph object to keep only components with clustered nodes}
\usage{
filter_igraph_hotnet(ig)
}
\arguments{
\item{ig}{An igraph object.}
}
\value{
A filtered igraph object.
}
\description{
Filter igraph object to keep only components with clustered nodes
}
