% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_ldak_gene_elbow}
\alias{filter_ldak_gene_elbow}
\title{Filter LDAK results based on an elbow point}
\usage{
filter_ldak_gene_elbow(df, keep = NULL)
}
\arguments{
\item{df}{A data frame containing LDAK results with a 'score' column.}

\item{keep}{Optional character vector of genes to keep regardless of score.}
}
\value{
A filtered data frame.
}
\description{
Filter LDAK results based on an elbow point
}
