% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_ldak_gene_fdr}
\alias{filter_ldak_gene_fdr}
\title{Filter LDAK gene scores based on FDR threshold}
\usage{
filter_ldak_gene_fdr(df, pval.col, by.col = NULL, threshold = 1, keep = NULL)
}
\arguments{
\item{df}{Data frame of LDAK scores.}

\item{pval.col}{P-value column for FDR.}

\item{by.col}{Grouping columns.}

\item{threshold}{FDR threshold.}

\item{keep}{Genes to keep.}
}
\value{
Filtered data frame.
}
\description{
Filter LDAK gene scores based on FDR threshold
}
