% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_ldak_gene_pvalue}
\alias{filter_ldak_gene_pvalue}
\title{Filter LDAK gene scores based on p-value threshold}
\usage{
filter_ldak_gene_pvalue(df, pval.col, threshold = 1, keep = NULL)
}
\arguments{
\item{df}{Data frame of LDAK scores.}

\item{pval.col}{P-value column name.}

\item{threshold}{Threshold value.}

\item{keep}{Genes to keep.}
}
\value{
Filtered data frame.
}
\description{
Filter LDAK gene scores based on p-value threshold
}
