% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_ldak_gene_SD}
\alias{filter_ldak_gene_SD}
\title{Filter LDAK gene scores based on SD values}
\usage{
filter_ldak_gene_SD(df, keep = NULL)
}
\arguments{
\item{df}{Data frame of LDAK scores.}

\item{keep}{Genes to keep.}
}
\value{
Filtered data frame.
}
\description{
Filter LDAK gene scores based on SD values
}
