% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_ldak_gene_type}
\alias{filter_ldak_gene_type}
\title{Filter LDAK gene scores based on gene types}
\usage{
filter_ldak_gene_type(df, gtf, types = c("protein_coding"), keep = NULL)
}
\arguments{
\item{df}{Data frame of LDAK scores.}

\item{gtf}{GTF annotation.}

\item{types}{Gene types to keep.}

\item{keep}{Genes to keep.}
}
\value{
Filtered data frame.
}
\description{
Filter LDAK gene scores based on gene types
}
