% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{find_na}
\alias{find_na}
\title{Find rows with NA values in any column}
\usage{
find_na(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data frame containing only rows with at least one NA value.
}
\description{
Find rows with NA values in any column
}
