% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Internal helper to determine and create a cache directory}
\usage{
get_cache_dir(sub_dir = NULL, cache_dir = NULL)
}
\arguments{
\item{sub_dir}{Optional sub-directory name to create within the cache.}

\item{cache_dir}{Optional user-provided base cache directory.}
}
\value{
The absolute path to the (sub-)cache directory.
}
\description{
Internal helper to determine and create a cache directory
}
\keyword{internal}
