% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{get_data_refs_dir}
\alias{get_data_refs_dir}
\title{Internal helper to determine and create a data_refs directory}
\usage{
get_data_refs_dir(data_refs_dir = NULL)
}
\arguments{
\item{data_refs_dir}{Optional user-provided directory.}
}
\value{
The absolute path to the data_refs directory.
}
\description{
Internal helper to determine and create a data_refs directory
}
\keyword{internal}
