% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_access.R
\name{get_hic}
\alias{get_hic}
\title{Get both PP and PO HiC data by tissue type}
\usage{
get_hic(tissue_type = NULL, cache_dir = NULL)
}
\arguments{
\item{tissue_type}{Character vectors of tissue types to retain (e.g., "Esophagus", "Gastric"). If NULL, returns all tissues.}

\item{cache_dir}{Root directory containing HiC resources.}
}
\value{
A named list containing 'PP' and 'PO' data frames.
}
\description{
Get both PP and PO HiC data by tissue type
}
