% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_access.R
\name{get_hic_by_tissue}
\alias{get_hic_by_tissue}
\title{Get HiC data by tissue type}
\usage{
get_hic_by_tissue(
  tissue_type = NULL,
  edge_type = c("PO", "PP"),
  cache_dir = NULL
)
}
\arguments{
\item{tissue_type}{Character vectors of tissue types to retain (e.g., "Esophagus", "Gastric"). If NULL, returns all tissues.}

\item{edge_type}{Character. Either "PO" (Promoter-Other) or "PP" (Promoter-Promoter)}

\item{cache_dir}{Root directory containing HiC resources.}
}
\value{
A data frame of the processed HiC interactions filtered to the requested tissue(s)
}
\description{
Get HiC data by tissue type
}
