% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEME_access.R
\name{get_jeme}
\alias{get_jeme}
\title{Get JEME data for specific tissues}
\usage{
get_jeme(
  key_column = NULL,
  value = NULL,
  method = c("lasso", "elasticnet"),
  simplified = TRUE,
  cache_dir = NULL
)
}
\arguments{
\item{key_column}{Character. The column name in the key_tissues_jeme dataset to match (e.g., "desc1"). If NULL (default), reads all tissues.}

\item{value}{Character vector. The values to match in the key_column.}

\item{method}{Character. Either "lasso" or "elasticnet"}

\item{simplified}{Logical. If TRUE, concatenates the resulting list into a single data frame}

\item{cache_dir}{Root directory containing the JEME resource directory.}
}
\value{
A list of data frames, or a single combined data frame if simplified=TRUE
}
\description{
Get JEME data for specific tissues
}
