% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEME_access.R
\name{get_jeme_by_tissue}
\alias{get_jeme_by_tissue}
\title{Get JEME data by tissue nfile}
\usage{
get_jeme_by_tissue(nfile, method = c("lasso", "elasticnet"), cache_dir = NULL)
}
\arguments{
\item{nfile}{Numeric or character. The tissue numeric file identifier (e.g., 77, 92)}

\item{method}{Character. Either "lasso" or "elasticnet"}

\item{cache_dir}{Root directory containing JEME and HiC resources.}
}
\value{
A data frame of the processed JEME interactions for that tissue
}
\description{
Get JEME data by tissue nfile
}
