% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ldak}
\alias{get_ldak}
\title{Load and process LDAK results}
\usage{
get_ldak(file, cohort = NULL)
}
\arguments{
\item{file}{Path to the LDAK results file in \code{.rds} format.}

\item{cohort}{Optional character string for GWAS cohort filtering.}
}
\value{
A tibble of LDAK results.
}
\description{
Load and process LDAK results
}
