% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_regulator_LDAK_scores}
\alias{get_regulator_LDAK_scores}
\title{Get LDAK scores for enhancers and promoters}
\usage{
get_regulator_LDAK_scores(param, enhancers, verbose = TRUE)
}
\arguments{
\item{param}{A list of parameters.}

\item{enhancers}{Character vector of enhancer IDs.}

\item{verbose}{Logical, whether to print progress.}
}
\value{
A tibble of combined LDAK scores.
}
\description{
Get LDAK scores for enhancers and promoters
}
