% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_access.R
\name{harmonize_hic_gtf}
\alias{harmonize_hic_gtf}
\title{Harmonize HiC networks using preprocessed alias link}
\usage{
harmonize_hic_gtf(hic, alias_link_file, verbose = TRUE, rm_na = FALSE)
}
\arguments{
\item{hic}{A list containing PO and PP networks}

\item{alias_link_file}{Path to alias link RDS}

\item{verbose}{Logical.}

\item{rm_na}{Logical.}
}
\value{
Harmonized HiC networks
}
\description{
Harmonize HiC networks using preprocessed alias link
}
