% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEME_access.R
\name{harmonize_promoter_by_ensg}
\alias{harmonize_promoter_by_ensg}
\title{Harmonize gene symbols by Ensembl ID with GENCODE annotation, to match those in GTEx V10}
\usage{
harmonize_promoter_by_ensg(df, gtf = NULL, verbose = TRUE)
}
\arguments{
\item{df}{A data frame containing at least \code{ENSG} and \code{promoter} columns.}

\item{gtf}{A tibble/dataframe containing GENCODE annotation. If NULL (default),
attempts to read the cached GTEx GENCODE GTF.}

\item{verbose}{Logical, whether to print progress messages (default = TRUE)}
}
\value{
A data frame with an added \code{promoter_harmonized} column.
}
\description{
Harmonize gene symbols by Ensembl ID with GENCODE annotation, to match those in GTEx V10
}
