% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{hotnet_cluster_df}
\alias{hotnet_cluster_df}
\title{Convert hotnet clusters list to a data frame}
\usage{
hotnet_cluster_df(res_net)
}
\arguments{
\item{res_net}{A list.}
}
\value{
A data frame.
}
\description{
Convert hotnet clusters list to a data frame
}
