% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldak_utils.R
\name{jeme_promoter_to_loc}
\alias{jeme_promoter_to_loc}
\title{Convert JEME promoter nodes to LDAK .loc format}
\usage{
jeme_promoter_to_loc(
  promoters,
  output_file = NULL,
  annotation_src = c("NCBI", "jeme"),
  ncbi_tbl = NULL
)
}
\arguments{
\item{promoters}{A data frame of JEME promoter nodes, as returned by \code{extract_jeme_nodes()$promoters}.}

\item{output_file}{Path to the output .loc file. If NULL, defaults to JEME subdirectory in data reference.}

\item{annotation_src}{Character. Either "NCBI" (use NCBI feature table for ranges) or "jeme" (use coordinates in JEME data).}

\item{ncbi_tbl}{A data frame from \code{NCBI_feature_table()}, required if \code{annotation_src == "NCBI"}.}
}
\value{
A data frame of the formatted LDAK boundary data.
}
\description{
Convert JEME promoter nodes to LDAK .loc format
}
