% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{key_tissues_jeme}
\alias{key_tissues_jeme}
\title{JEME Tissue Mapping Key}
\format{
A data frame with information on JEME tissues:
\describe{
\item{nfile}{Numeric file identifier used in JEME filenames}
\item{file}{Filename of the JEME dataset}
\item{cat1}{Main category/system}
\item{cat2}{Sub-category/specific tissue}
\item{desc1}{Long description of the tissue}
\item{desc2}{Additional description or source info}
\item{Epigenome mnemonic}{Roadmap epigenome mnemonic}
\item{input}{Input identifier}
\item{input_name}{Input name}
\item{Sample group}{Sample grouping}
\item{Sample name}{Sample description}
\item{tiss}{Tissue identifier}
}
}
\source{
\url{https://labs.sbpdiscovery.org/centerandlabs/cancercenter/YipLab/jeme/}
}
\usage{
data(key_tissues_jeme)
}
\description{
A dataset containing mapping information for JEME (Joint Effects of Mutations and Enhancers) results.
This key helps associate tissue numeric identifiers with descriptive names and categories.
}
\keyword{datasets}
