% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_access.R
\name{load_hic}
\alias{load_hic}
\title{Deprecated: Load HiC networks}
\usage{
load_hic(param, verbose = TRUE, rm_na = FALSE)
}
\arguments{
\item{param}{Parameters list.}

\item{verbose}{Logical.}

\item{rm_na}{Logical.}
}
\description{
This function is deprecated. Please use \code{\link{get_hic}} instead.
}
