% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{load_hotnet_input}
\alias{load_hotnet_input}
\title{Load hotnet network input files}
\usage{
load_hotnet_input(net_prefix = "network_1", data_dir)
}
\arguments{
\item{net_prefix}{String specifying the network prefix (default: "network_1")}

\item{data_dir}{Path to data directory (required)}
}
\value{
A data frame containing network edges with columns from and to nodes
}
\description{
Load hotnet network input files
}
