% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{load_hotnet_results}
\alias{load_hotnet_results}
\title{Load hotnet results from clusters file}
\usage{
load_hotnet_results(network = "network_1", result_dir, exclude_single = FALSE)
}
\arguments{
\item{network}{String specifying the network name (default: "network_1")}

\item{result_dir}{Path to results directory (required)}

\item{exclude_single}{Logical indicating whether to exclude single-node clusters (default: FALSE)}
}
\value{
A list.
}
\description{
Load hotnet results from clusters file
}
