% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEME_access.R
\name{load_jeme}
\alias{load_jeme}
\title{Deprecated: Load JEME network}
\usage{
load_jeme(param, verbose = TRUE)
}
\arguments{
\item{param}{Parameters list.}

\item{verbose}{Logical.}
}
\description{
This function is deprecated. Please use \code{\link{get_jeme}} instead.
}
