% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge values of a column by group}
\usage{
merge_rows(df, group_cols, merge_col, sep = ";")
}
\arguments{
\item{df}{A data frame.}

\item{group_cols}{Character vector of column names to group by.}

\item{merge_col}{Name of the column to merge.}

\item{sep}{Separator to use for merging (default = ";").}
}
\value{
A data frame with merged rows.
}
\description{
Merge values of a column by group
}
