% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{merge_with_order}
\alias{merge_with_order}
\title{Merge preserving the original order if sort==FALSE}
\usage{
merge_with_order(x, y, ..., sort = TRUE)
}
\description{
Merge preserving the original order if sort==FALSE
}
\keyword{internal}
