% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nlog10}
\alias{nlog10}
\title{Calculate negative log base 10, handling infinite values}
\usage{
nlog10(x, max_value = 1000)
}
\arguments{
\item{x}{A numeric vector.}

\item{max_value}{Value to substitute for Inf (default = 1000).}
}
\value{
Negative log10 values.
}
\description{
Calculate negative log base 10, handling infinite values
}
