% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nlog2}
\alias{nlog2}
\title{Calculate negative log base 2}
\usage{
nlog2(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Arguments passed to nlog10.}
}
\value{
Negative log2 values.
}
\description{
Calculate negative log base 2
}
