% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plot_subgraphs_to_pdf}
\alias{plot_subgraphs_to_pdf}
\title{Plot each connected component of an igraph object to a multi-page PDF}
\usage{
plot_subgraphs_to_pdf(ig, score_attr = "score", pdf_file = "subgraphs.pdf")
}
\arguments{
\item{ig}{An \code{igraph} object.}

\item{score_attr}{Character string.}

\item{pdf_file}{Character string.}
}
\value{
No return value. A PDF file is created.
}
\description{
Plot each connected component of an igraph object to a multi-page PDF
}
