% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_subgraphs_to_pdf2}
\alias{plot_subgraphs_to_pdf2}
\title{Plot each connected component of an igraph object to a multi-page PDF}
\usage{
plot_subgraphs_to_pdf2(
  ig,
  score_attr = "score",
  pdf_file = "subgraphs.pdf",
  node_names = FALSE
)
}
\arguments{
\item{ig}{An igraph object.}

\item{score_attr}{Character string.}

\item{pdf_file}{Character string.}

\item{node_names}{Optional vector or logical.}
}
\value{
No return value.
}
\description{
Plot each connected component of an igraph object to a multi-page PDF
}
