% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{print_network_info}
\alias{print_network_info}
\title{Print summary information of an igraph network}
\usage{
print_network_info(ig, score_attr = "score", top_n = 6)
}
\arguments{
\item{ig}{An igraph object.}

\item{score_attr}{Character string.}

\item{top_n}{Number of top/bottom components to display.}
}
\value{
Invisibly returns a tibble with subgraph/component information.
}
\description{
Print summary information of an igraph network
}
