% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{promoter_loc}
\alias{promoter_loc}
\title{Extract promoter locations from JEME and HiC data}
\usage{
promoter_loc(jeme, hic, promoter_loc_file, verbose = TRUE)
}
\arguments{
\item{jeme}{A data frame with column Target}

\item{hic}{A list returned by load_hic() containing elements PO and PP}

\item{promoter_loc_file}{Path to LDAK promoter .loc file}

\item{verbose}{Logical, whether to print progress messages}
}
\value{
A data frame of promoter locations.
}
\description{
Extract promoter locations from JEME and HiC data
}
