% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_ARACNe}
\alias{read_ARACNe}
\title{Read and filter ARACNe network file}
\usage{
read_ARACNe(
  infile,
  MI_threshold = 0,
  pvalue_threshold = 1e-06,
  keep_genes = NULL
)
}
\arguments{
\item{infile}{Path to the ARACNe network file.}

\item{MI_threshold}{Minimum MI threshold.}

\item{pvalue_threshold}{Maximum p-value threshold.}

\item{keep_genes}{Optional vector of genes to keep.}
}
\value{
A tibble of filtered ARACNe edges.
}
\description{
Read and filter ARACNe network file
}
