% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_GTEx.R
\name{read_gtex_gencode_gtf}
\alias{read_gtex_gencode_gtf}
\title{Read and clean GTEx GENCODE annotation from cache}
\usage{
read_gtex_gencode_gtf(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Cache directory.}
}
\value{
A data frame of cleaned GENCODE genes.
}
\description{
Read and clean GTEx GENCODE annotation from cache
}
\keyword{internal}
