% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_ldak_gene_score}
\alias{read_ldak_gene_score}
\title{Read LDAK gene score from file}
\usage{
read_ldak_gene_score(flk, cohort = NULL, file)
}
\arguments{
\item{flk}{Flank size.}

\item{cohort}{GWAS cohort.}

\item{file}{Path to RDS file.}
}
\value{
A filtered data frame of LDAK scores.
}
\description{
Read LDAK gene score from file
}
