% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{save_hotnet_network_files}
\alias{save_hotnet_network_files}
\title{Save HotNet network files}
\usage{
save_hotnet_network_files(obj, network_name, out_dir = "./data")
}
\arguments{
\item{obj}{A list returned by \code{igraph_to_hotnet()}.}

\item{network_name}{Character string.}

\item{out_dir}{Character string.}
}
\value{
Invisibly returns a list of file paths.
}
\description{
Save HotNet network files
}
