% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{select_components_by_nodes}
\alias{select_components_by_nodes}
\title{Select subgraphs containing specified nodes}
\usage{
select_components_by_nodes(ig, nodes)
}
\arguments{
\item{ig}{An \code{igraph} object.}

\item{nodes}{A character vector of node names.}
}
\value{
An \code{igraph} object.
}
\description{
Select subgraphs containing specified nodes
}
