% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{select_rows_with_semicolon}
\alias{select_rows_with_semicolon}
\title{Select rows based on presence of semicolons in a column}
\usage{
select_rows_with_semicolon(df, col, invert = FALSE)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{col}{A column to check for semicolons (unquoted).}

\item{invert}{Logical (default = FALSE).}
}
\value{
A tibble containing only the rows that satisfy the condition.
}
\description{
Select rows based on presence of semicolons in a column
}
