% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{split_columns}
\alias{split_columns}
\title{Split columns by semicolon into multiple rows}
\usage{
split_columns(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{One or more unquoted column names (tidy-select) to split.}
}
\value{
A data frame where columns are split into multiple rows.
}
\description{
Split columns by semicolon into multiple rows
}
