% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{subgraph_info_FDR}
\alias{subgraph_info_FDR}
\title{Summarize subgraph/component information in an igraph object (using FDR)}
\usage{
subgraph_info_FDR(ig, score_attr = "score")
}
\arguments{
\item{ig}{An \code{igraph} object.}

\item{score_attr}{Character string.}
}
\value{
A tibble with columns.
}
\description{
Summarize subgraph/component information in an igraph object (using FDR)
}
