% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEME_access.R
\name{summary_harmonized_jeme}
\alias{summary_harmonized_jeme}
\title{Deprecated: Summarize harmonized JEME results}
\usage{
summary_harmonized_jeme(jeme_harmonized, max_print = 10, verbose = TRUE)
}
\arguments{
\item{jeme_harmonized}{A tibble returned by \code{harmonize_jeme_genecode()} or \code{harmonize_promoter_by_ensg()}}

\item{max_print}{Maximum items to print.}

\item{verbose}{Logical.}
}
\value{
A summary list.
}
\description{
This function is deprecated.
}
