% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tf_25}
\alias{tf_25}
\title{Selection of 25 Transcription Factors}
\format{
A data frame with one column:
\describe{
\item{gene_symbol}{HUGO Gene Symbol}
}
}
\usage{
data(tf_25)
}
\description{
A prioritized list of 25 transcription factors found in \code{data-raw/TF/TF25.tsv}.
}
\keyword{datasets}
